refactor(users): align architecture with clean principles and simplify models
