+++ Users/Infrastructure/User.cs
@@ -2,12 +2,9 @@
 
 public class User
 {
-    public Guid Id { get; private set; } = Guid.NewGuid();
+    public Guid Id { get; private set; }
     public string Email { get; set; } = string.Empty;
     public string PasswordHash { get; set; } = string.Empty;
-
-    private readonly List<Guid> _profileIds = new();
-    public IReadOnlyList<Guid> ProfileIds => _profileIds.AsReadOnly();
 
     private User() { }
 
@@ -15,16 +12,5 @@ public class User
     {
         Id = Guid.NewGuid();
         Email = email;
-    }
-
-    public void AddProfile(Guid profileId)
-    {
-        if (!_profileIds.Contains(profileId))
-            _profileIds.Add(profileId);
-    }
-
-    public void RemoveProfile(Guid profileId)
-    {
-        _profileIds.Remove(profileId);
     }
 }
